fix: blockchain verifier end-to-end working + port 8001 + web3 v7
- Upgrade web3 v5->v7.14.1 (Python 3.12 compatible, no pkg_resources)
- Fix _load_abi() to read from hardhat AuditLog.json artifact (not empty deployment.json)
- Add store_log_on_chain() call in ingest_log() via background task
- Add update_log() to mongo_client for patching tx_hash/block_number
- Add _store_on_chain() using asyncio.run_coroutine_threadsafe for async DB update from thread
- Fix storeLog() call: add missing logId (MongoDB ObjectId) as first arg (contract has 6 params)
- Fix get_log_from_chain() to accept str logId not int index (contract uses getLog(string))
- Fix verify.py to pass log_id string to getLog() instead of chain_index int
- Fix compute_log_hash() to use canonical _HASH_FIELDS list (stable hash pre/post MongoDB save)
- Switch backend port from 8000 to 8001 (Windows Hyper-V reserves 8000)
- Update vite.config.js proxy target to http://localhost:8001
- Update simulate_events.py API_URL to port 8001
- Capture uvicorn event loop in startup and inject into routes.logs._main_loop
- Update README.md with Hardhat startup steps, port 8001 note, verifier docs, troubleshooting
- Regenerate all 3 PDFs (Project Guide, API Docs, HowTo)

+++ backend/models/train_models.py
@@ -236,7 +236,7 @@ def train_models():
     
     # Evaluate on full dataset
     predictions = iso_forest.predict(X_scaled)
-    anomaly_rate = (predictions == -1).mean()
+    anomaly_rate = float((predictions == -1).mean())
     print(f"      Saved: anomaly_model.pkl")
     print(f"      Anomaly rate on full dataset: {anomaly_rate:.2%}")
